refactor: :recycle: modified structure
