+++ src/models/layout.ts
@@ -1,0 +1,5 @@
+import { ReactElement } from "react";
+
+export interface LayoutProps {
+  children: ReactElement;
+}
